refactor:#265 change password tpye파일명 변경
