+++ src/app/(auth)/change-password/_components/change-password-funnel.tsx
@@ -2,7 +2,7 @@ import useFunnel from '@/hooks/use-funnel';
 import StepComplete from './step-complete';
 import StepNewPassword from './step-new-password';
 import StepCurrentPassword from './step-current-password';
-import { ChangePasswordFunnelStep, ChangePasswordFunnelType, StepNewPasswordType } from '../_types/funnel-type';
+import { ChangePasswordFunnelStep, ChangePasswordFunnelType, StepNewPasswordType } from '../_types/funnel.type';
 import { z } from 'zod';
 import formSchema from '@/app/schemas/form-schema.schema';
 import { memo } from 'react';
