+++ src/app/(auth)/change-password/_types/funnel.type.ts
