+++ src/app/(auth)/change-password/page.tsx
@@ -2,7 +2,7 @@
 
 import { useSearchParams } from 'next/navigation';
 import GlassBackground from '@/components/commons/glass-background';
-import { ChangePasswordFunnelStep } from './_types/funnel-type';
+import { ChangePasswordFunnelStep } from './_types/funnel.type';
 import ChangePasswordFunnel from './_components/change-password-funnel';
 
 const ResetPasswordPage = () => {
